+++ tests/scalers/memory/memory_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/joho/godotenv"
 	"github.com/stretchr/testify/assert"
+	"k8s.io/client-go/kubernetes"
 
 	. "github.com/kedacore/keda/v2/tests/helper"
 )
@@ -20,11 +21,25 @@ const (
 	testName = "memory-test"
 )
 
+var (
+	testNamespace          = fmt.Sprintf("%s-ns", testName)
+	deploymentName         = fmt.Sprintf("%s-deployment", testName)
+	scaledObjectName       = fmt.Sprintf("%s-so", testName)
+	workloadDeploymentName = fmt.Sprintf("%s-workload-deployment", testName)
+	minReplicas            = 0
+	maxReplicas            = 5
+	utilizationValue       = 45 // downScale value
+)
+
 type templateData struct {
 	TestNamespace    string
 	DeploymentName   string
 	ScaledObjectName string
 	UtilizationValue int32
+
+	MinReplicas            string
+	MaxReplicas            string
+	WorkloadDeploymentName string
 }
 
 const (
@@ -32,6 +47,8 @@ const (
 apiVersion: apps/v1
 kind: Deployment
 metadata:
+  labels:
+    deploy: {{.DeploymentName}}
   name: {{.DeploymentName}}
   namespace: {{.TestNamespace}}
 spec:
@@ -86,27 +103,84 @@ spec:
       value: "{{.UtilizationValue}}"
 `
 
+	scaledObjectTwoTriggerTemplate = `
+apiVersion: keda.sh/v1alpha1
+kind: ScaledObject
+metadata:
+  name: {{.ScaledObjectName}}
+  namespace: {{.TestNamespace}}
+  labels:
+    run: {{.DeploymentName}}
+spec:
+  advanced:
+    horizontalPodAutoscalerConfig:
+      behavior:
+        scaleDown:
+          stabilizationWindowSeconds: 1
+  pollingInterval: 1
+  scaleTargetRef:
+    name: {{.DeploymentName}}
+  minReplicaCount: {{.MinReplicas}}
+  maxReplicaCount: {{.MaxReplicas}}
+  cooldownPeriod: 1
+  triggers:
+  - type: memory
+    metadata:
+      type: Utilization
+      value: "{{.UtilizationValue}}"
+  - type: kubernetes-workload
+    metadata:
+      podSelector: 'pod={{.WorkloadDeploymentName}}'
+      value: '1'
+`
+
+	workloadDeploymentTemplate = `apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: {{.WorkloadDeploymentName}}
+  namespace: {{.TestNamespace}}
+  labels:
+    deploy: {{.WorkloadDeploymentName}}
+spec:
+  replicas: 0
+  selector:
+    matchLabels:
+      pod: {{.WorkloadDeploymentName}}
+  template:
+    metadata:
+      labels:
+        pod: {{.WorkloadDeploymentName}}
+    spec:
+      containers:
+        - name: nginx
+          image: 'nginxinc/nginx-unprivileged'`
+
 	scaleUpValue   = 1
 	scaleDownValue = 45
 )
 
 func TestMemoryScaler(t *testing.T) {
-	testNamespace := fmt.Sprintf("%s-ns", testName)
-	deploymentName := fmt.Sprintf("%s-deployment", testName)
-	scaledObjectName := fmt.Sprintf("%s-so", testName)
-
 	// Create kubernetes resources
 	kc := GetKubernetesClient(t)
-	data, templates := getTemplateData(testNamespace, deploymentName, scaledObjectName, scaleUpValue)
+	data, templates := getTemplateData()
 
-	CreateKubernetesResources(t, kc, testNamespace, data, []Template{{Name: "deploymentTemplate", Config: deploymentTemplate}})
+	CreateKubernetesResources(t, kc, testNamespace, data, templates)
 
+	scaleOut(t, kc, data)
+	scaleToZero(t, kc, data)
+
+	// cleanup
+	DeleteKubernetesResources(t, kc, testNamespace, data, templates)
+}
+
+func scaleOut(t *testing.T, kc *kubernetes.Clientset, data templateData) {
 	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 1, 60, 1),
 		"Replica count should start out as 1")
 
 	t.Log("--- testing scale out ---")
 	t.Log("--- applying scaled object with scaled up utilization ---")
 
+	data.UtilizationValue = int32(scaleUpValue)
 	KubectlApplyMultipleWithTemplate(t, data, []Template{{Name: "scaledObjectTemplate", Config: scaledObjectTemplate}})
 
 	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 2, 180, 1),
@@ -115,24 +189,53 @@ func TestMemoryScaler(t *testing.T) {
 	t.Log("--- testing scale in ---")
 	t.Log("--- applying scaled object with scaled down utilization ---")
 
-	data, _ = getTemplateData(testNamespace, deploymentName, scaledObjectName, scaleDownValue)
+	data.UtilizationValue = int32(scaleDownValue)
 	KubectlApplyMultipleWithTemplate(t, data, []Template{{Name: "scaledObjectTemplate", Config: scaledObjectTemplate}})
 
 	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 1, 180, 1),
 		"Replica count should be 1 in next 3 minutes")
-
-	// cleanup
-	DeleteKubernetesResources(t, kc, testNamespace, data, templates)
 }
 
-func getTemplateData(testNamespace string, deploymentName string, scaledObjectName string, utilizationValue int32) (templateData, []Template) {
+func scaleToZero(t *testing.T, kc *kubernetes.Clientset, data templateData) {
+	t.Log("--- testing scale to zero ---")
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 1, 60, 1),
+		"Replica count should be 1")
+
+	AssertReplicaCountNotChangeDuringTimePeriod(t, kc, deploymentName, testNamespace, 1, 60)
+
+	// replica count is 1 without scaleToZero metadata field
+
+	KubectlApplyWithTemplate(t, data, "scaledObjectTwoTriggerTemplate", scaledObjectTwoTriggerTemplate)
+
+	// expect replica count to drop to 0 after updating SO with scaleToZero
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 0, 60, 1),
+		"Replica count should be 0")
+
+	// scale external trigger out (expect replicas scale out)
+	KubernetesScaleDeployment(t, kc, workloadDeploymentName, int64(maxReplicas), testNamespace)
+
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, maxReplicas, 60, 1),
+		"Replica count should be %v", maxReplicas)
+
+	// scale external trigger in (expect replicas back to 0 -- external trigger not active)
+	KubernetesScaleDeployment(t, kc, workloadDeploymentName, int64(minReplicas), testNamespace)
+
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, minReplicas, 60, 1),
+		"Replica count should be %v", minReplicas)
+}
+
+func getTemplateData() (templateData, []Template) {
 	return templateData{
-			TestNamespace:    testNamespace,
-			DeploymentName:   deploymentName,
-			ScaledObjectName: scaledObjectName,
-			UtilizationValue: utilizationValue,
+			TestNamespace:          testNamespace,
+			DeploymentName:         deploymentName,
+			ScaledObjectName:       scaledObjectName,
+			UtilizationValue:       int32(utilizationValue),
+			MinReplicas:            fmt.Sprintf("%v", minReplicas),
+			MaxReplicas:            fmt.Sprintf("%v", maxReplicas),
+			WorkloadDeploymentName: workloadDeploymentName,
 		}, []Template{
 			{Name: "deploymentTemplate", Config: deploymentTemplate},
 			{Name: "scaledObjectTemplate", Config: scaledObjectTemplate},
+			{Name: "workloadDeploymentTemplate", Config: workloadDeploymentTemplate},
 		}
 }